requirements.txt
flask
flask-socketio
pymongo
python-dotenv
flask-bcrypt
flask-login
eventlet
